extract last comments component
