+++ hugo/package.json
@@ -33,9 +33,9 @@
     "lodash": "^4.17.11",
     "lozad": "^1.9.0",
     "mark.js": "^8.11.1",
-    "node-sass": "^4.11.0",
+    "node-sass": "^4.12.0",
     "popmotion-pose": "^3.4.8",
-    "preact": "^8.4.2",
+    "preact": "^10.0.0-beta.1",
     "quicklink": "^1.0.0",
     "sass-loader": "^7.1.0",
     "scriptjs": "^2.5.9",
@@ -49,7 +49,7 @@
     "webpack": "^4.30.0"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/plugin-transform-react-jsx": "^7.3.0",
     "browser-sync": "^2.26.5",
